pyproject.toml
[project]
name = "literature-report-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=1.0.0",
  "langgraph-prebuilt>=1.0.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.1.0",
  "python-dotenv>=1.0.0",
  "requests>=2.0.0",
  "httpx>=0.28.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.0.0",
  "feedparser>=6.0.0",
  "pypdf>=4.0.0",
  "python-multipart>=0.0.20",
  "sentence-transformers>=3.0.0",
  "faiss-cpu>=1.7.0",
  "numpy>=1.26.0",
  "cohere>=5.0.0",
  "rank-bm25>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.uvicorn]
factory = false
